requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
boto3==1.24.19
botocore==1.27.19
certifi==2022.5.18.1
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangorestframework-stubs==1.7.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
msgpack==1.0.4
mypy==0.950
mypy-extensions==0.4.3
Pillow==9.1.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
toml==0.10.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
types-Markdown==3.4.0
types-pytz==2021.3.8
types-PyYAML==6.0.7
types-requests==2.28.3
types-urllib3==1.26.16
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
